Moved each validation to own class
